add manage team menu option
