+++ node-cli/projects.js
@@ -15,50 +15,3 @@ exports.getProjects = async () => {
   output.header("MY PROJECTS:");
   output.result(JSON.stringify(projects, null, 2));
 };
-
-
-const Choices = {
-  ShowAllProjects: "Show all of my projects",
-  ManageTeam: "Manage my team",
-  ReturnMainMenu: "Return to main menu",
-  LogOut: "Log out / Quit",
-};
-
-async function projectMenu() {
-  try {
-    const answers = await inquirer.prompt({
-      type: "rawlist",
-      name: "mainMenu",
-      message: "What would you like to do?",
-      choices: [...Object.values(Choices), new inquirer.Separator()],
-    });
-
-    switch (answers.projectMenu) {
-      case Choices.ShowAllProjects: {
-        await getProjects();
-        return projectMenu();
-      }
-      case Choices.ManageTeam: {
-        return team.manageTeamMenu();
-      }
-      case Choices.ReturnMainMenu: {
-        return main.mainMenu();
-      }
-      case Choices.LogOut: {
-        const loggedOut = await users.logOut();
-        if (!loggedOut) {
-          output.error("Error logging out");
-        } else output.result("You have been logged out. Use Ctrl-C to quit.");
-        return;
-      }
-      default: {
-        return projectMenu();
-      }
-    }
-  } catch (err) {
-    output.error(err);
-    return;
-  }
-}
-
-exports.projectMenu = projectMenu;
